third_party/WebKit/LayoutTests/bluetooth/characteristic/writeValue/value-too-long.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['generic_access']}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('generic_access'))
    .then(service => service.getCharacteristic('gap.device_name'))
    .then(characteristic => {
      return assert_promise_rejects_with_message(
        characteristic.writeValue(new Uint8Array(513 /* length */)),
        new DOMException(
          'Value can\'t exceed 512 bytes.', 'InvalidModificationError'),
        'Value passed was too long.');
    });
}, 'Trying to write more than 512 bytes should return an error.');
</script>
